Use a more AOT friendly method for json deserialization
The existing method in general works, but throws an IL1026 warning while trimming. Theres no universal way in a generic to generate a per enum deserialization scheme.

However, since the keys are always strings, it is possible to deserialize into a Dictionary<string, string>. That dictionary can then be easily mapped into the typed enums.

NS2.0 is missing the fast enum parse function, so the code is conditionally compiled to NS2.0 and the other targets